requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==3.2.12
djangorestframework==3.13.1
django-redis==5.2.0
djongo==1.3.6
django_celery_beat == 1.1.1
idna==3.3
gunicorn==19.9.0
marshmallow==3.15.0
packaging==21.3
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
redis==4.2.2
six==1.16.0
sqlparse==0.2.4
twilio==7.8.0
typing_extensions==4.1.1
urllib3==1.26.9
dnspython==2.2.1
fuzzywuzzy==0.18.0
python_Levenshtein==0.12.2
psycopg2-binary == 2.9.3
dj-database-url==0.5.0
websockets == 10.1
asyncio == 3.4.3
asgiref == 3.5.0
celery == 4.4.7